+++ src/Speed/Indep/Src/Frontend/MenuScreens/Safehouse/career/uiRapSheetCTS.hpp
@@ -41,7 +41,7 @@ class uiRapSheetCTS : public ArrayScrollerMenu {
   public:
     uiRapSheetCTS(ScreenConstructorData *sd);
     ~uiRapSheetCTS() override {};
-    void NotificationMessage(unsigned long msg, FEObject *pobj, unsigned long param1, unsigned long param2) override;
+    void NotificationMessage(u32 msg, FEObject *pobj, u32 param1, u32 param2) override;
     void RefreshHeader() override;
 
   private:
